tuning null -> ... analysis (dereference is a total bottom)
If there is a dereference of a parameter, then the approximation is Bottom. A tricky change in engine: if `foo(param)` is analyzed and there is a call inside foo to interface `myInterface.someMethod(param)` then we optimistically assume then we optimistically assume that this call will not produce exceptions.